improve package listing display
This tries to make the package listing/display a little more user
friendly.  The primary way it does this is by separating a package name
like

bro/user/package into 'package' and 'user' so they can be displayed
separately.  The import process could probably do this step a little
better.

This still needs a bit of work but it's a start.  I couldn't get the
fontawesome stuff to work and couldn't figure out which version of
'primer' is present.