Redirect away from noisy package/tag view URLs.
Crawlers picked up on the fact that there can be noise after /packages/view/<ID>
and /tags/view/<ID> URL paths, and our setup so far preserved that noise when
rendering the view. We now strip it. I'm not sure this is the right fix -- seems
there should be something more "high-level" that redirects.