Drop markdown helper plugin, implement it ourselves
This is future-proofing for a move to CakePHP 5.x. The Tanuck plugin is
broken for CakePHP 4.x, and the one working for (kitzberger) doesn't
support 5.x. Better to just remove all of it and replace it with
something that hopefully won't go out of date. The plugin was very
simple in the end and wasn't really necessary as a dependency.